Split GraphModelBase into own files
